backend/requirements.txt
django==5.2.7
asgiref==3.10.0  
sqlparse==0.5.3 
tzdata==2025.2  
pillow==12.0.0
djangorestframework==3.16.1
django-filter==25.2
markdown==3.9
djangorestframework-simplejwt==5.5.1 
pyjwt==2.10.1
django-cors-headers==4.9.0
django-environ==0.12.0
psycopg2-binary==2.9.11
geographiclib==2.1 
geopy==2.4.1 